+++ client/web/src/enterprise/batches/backend.ts
@@ -102,6 +102,27 @@ export const queryBatchChangeBatchSpecs = (id: Scalars['ID']) => ({
         })
     )
 
+const PARTIAL_BATCH_WORKSPACE_FILE_FIELDS = gql`
+    fragment PartialBatchSpecWorkspaceFileFields on BatchSpecWorkspaceFile {
+        id
+        name
+        binary
+        byteSize
+    }
+`
+
+const BATCH_WORKSPACE_FILE_FIELDS = gql`
+    fragment BatchSpecWorkspaceFileFields on BatchSpecWorkspaceFile {
+        ...PartialBatchSpecWorkspaceFileFields
+        highlight(disableTimeout: false) {
+            aborted
+            html
+        }
+    }
+
+    ${PARTIAL_BATCH_WORKSPACE_FILE_FIELDS}
+`
+
 const BATCH_SPEC_LIST_FIELDS_FRAGMENT = gql`
     fragment BatchSpecListFields on BatchSpec {
         __typename
@@ -112,6 +133,7 @@ const BATCH_SPEC_LIST_FIELDS_FRAGMENT = gql`
         createdAt
         source
         description {
+            __typename
             name
         }
         namespace {
@@ -122,7 +144,19 @@ const BATCH_SPEC_LIST_FIELDS_FRAGMENT = gql`
             username
         }
         originalInput
+        files {
+            totalCount
+            pageInfo {
+                endCursor
+                hasNextPage
+            }
+            nodes {
+                ...PartialBatchSpecWorkspaceFileFields
+            }
+        }
     }
+
+    ${PARTIAL_BATCH_WORKSPACE_FILE_FIELDS}
 `
 
 const BATCH_SPEC_LIST_CONNECTION_FIELDS = gql`
@@ -139,3 +173,28 @@ const BATCH_SPEC_LIST_CONNECTION_FIELDS = gql`
 
     ${BATCH_SPEC_LIST_FIELDS_FRAGMENT}
 `
+
+export const BATCH_SPEC_WORKSPACE_FILE = gql`
+    query BatchSpecWorkspaceFile($id: ID!) {
+        node(id: $id) {
+            ... on BatchSpecWorkspaceFile {
+                ...BatchSpecWorkspaceFileFields
+            }
+        }
+    }
+
+    ${BATCH_WORKSPACE_FILE_FIELDS}
+`
+
+const BATCH_CHANGE_FILE_BASE_URL = '/.api/files/batch-changes'
+
+export const generateFileDownloadLink = async (specId: string, fileId: string): Promise<string> => {
+    const url = `${BATCH_CHANGE_FILE_BASE_URL}/${specId}/${fileId}`
+    const file = await fetch(url, {
+        headers: {
+            ...window.context.xhrHeaders,
+        },
+    })
+    const fileBlob = await file.blob()
+    return URL.createObjectURL(fileBlob)
+}