+++ client/web/src/enterprise/batches/detail/BatchChangeDetailsPage.mock.ts
@@ -71,6 +71,31 @@ export const MOCK_BATCH_CHANGE: BatchChangeFields = {
             nodes: [],
             totalCount: 0,
         },
+        description: {
+            __typename: 'BatchChangeDescription',
+            name: 'Spec Description',
+        },
+        files: {
+            totalCount: 2,
+            pageInfo: {
+                endCursor: null,
+                hasNextPage: false,
+            },
+            nodes: [
+                {
+                    id: 'random-id',
+                    name: 'test.txt',
+                    binary: false,
+                    byteSize: 74,
+                },
+                {
+                    id: 'random-id-2',
+                    name: 'src-cli',
+                    binary: true,
+                    byteSize: 75,
+                },
+            ],
+        },
     },
     batchSpecs: {
         nodes: [{ state: BatchSpecState.COMPLETED }],